+++ janitor-lang/src/main/java/com/eischet/janitor/env/JSetClass.java
@@ -1,14 +1,14 @@
 package com.eischet.janitor.env;
 
 import com.eischet.janitor.api.JanitorScriptProcess;
-import com.eischet.janitor.api.types.functions.JCallArgs;
 import com.eischet.janitor.api.errors.runtime.JanitorRuntimeException;
-import com.eischet.janitor.api.types.wrapped.JanitorWrapper;
-import com.eischet.janitor.api.types.*;
+import com.eischet.janitor.api.types.JanitorObject;
 import com.eischet.janitor.api.types.builtin.JBool;
 import com.eischet.janitor.api.types.builtin.JInt;
 import com.eischet.janitor.api.types.builtin.JList;
 import com.eischet.janitor.api.types.builtin.JSet;
+import com.eischet.janitor.api.types.functions.JCallArgs;
+import com.eischet.janitor.api.types.wrapped.JanitorWrapper;
 
 import java.util.Set;
 
@@ -30,6 +30,12 @@ public class JSetClass {
     public static JBool __contains(final JanitorWrapper<Set<JanitorObject>> _self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         final JSet self = ((JSet) _self);
         return JBool.of(self.contains(arguments.require(1).get(0)));
+    }
+
+    public static JSet __toSet(final JanitorWrapper<Set<JanitorObject>> _self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
+        final JSet self = ((JSet) _self);
+        arguments.require(0);
+        return process.getEnvironment().getBuiltinTypes().set(self.janitorGetHostValue().stream());
     }
 
     public static JList __toList(final JanitorWrapper<Set<JanitorObject>> _self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {